commands: Default to remote checks and introduce `--local-only` flag
When a remote dependency is specified in a zeitgeist config file, we
should always check for it, unless we explicitly request a local check.

We also add the `--local-only` flag to clarify that a local check does
not also check remotes.

The following flags have been marked deprecated as a result:
- `--local`
- `--remote`

Signed-off-by: Stephen Augustus <foo@auggie.dev>