Fix local checks logic
I recently noticed that db0fa72, by switching the logic around
local/remote checking, inadvertently made zeitgeist bypass local checks
by default (!).

This is a pretty major bug, as we lose half of the functionality
promised in the README (ensuring consistencies are kept up-to-date
within a project).

This commit fixes the logic to:
- Always run local checks
- Run remote checks, unless the --local-only flag is passed